+++ packages/duckdb-server-rust/src/lib.rs
@@ -1,0 +1,13 @@
+mod app;
+mod bundle;
+mod cache;
+mod db;
+mod interfaces;
+mod query;
+mod websocket;
+
+pub use app::app;
+pub use cache::{get_key, retrieve};
+pub use db::{ConnectionPool, Database};
+pub use interfaces::{AppError, AppState, Command, QueryParams, QueryResponse};
+pub use query::handle;
