+++ packages/duckdb-server-rust/src/main.rs
@@ -1,81 +1,18 @@
 use anyhow::Result;
-use axum::{
-    extract::{Query, State, WebSocketUpgrade},
-    http::Method,
-    response::Json,
-    routing::get,
-    Router,
-};
 use axum_server::tls_rustls::RustlsConfig;
 use listenfd::ListenFd;
 use std::net::TcpListener;
-use std::sync::Arc;
-use std::time::Duration;
 use std::{net::Ipv4Addr, net::SocketAddr, path::PathBuf};
 use tokio::net;
-use tokio::sync::Mutex;
-use tower_http::cors::{Any, CorsLayer};
-use tower_http::{compression::CompressionLayer, trace::TraceLayer};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
+mod app;
 mod bundle;
 mod cache;
 mod db;
 mod interfaces;
 mod query;
 mod websocket;
-
-use db::ConnectionPool;
-use interfaces::{AppError, AppState, QueryParams, QueryResponse};
-
-async fn handle_get(
-    State(state): State<Arc<AppState>>,
-    ws: Option<WebSocketUpgrade>,
-    Query(params): Query<QueryParams>,
-) -> Result<QueryResponse, AppError> {
-    if let Some(ws) = ws {
-        // WebSocket upgrade
-        Ok(QueryResponse::Response(
-            ws.on_upgrade(|socket| websocket::handle(socket, state)),
-        ))
-    } else {
-        // HTTP request
-        query::handle(&state, params).await
-    }
-}
-
-async fn handle_post(
-    State(state): State<Arc<AppState>>,
-    Json(params): Json<QueryParams>,
-) -> Result<QueryResponse, AppError> {
-    query::handle(&state, params).await
-}
-
-pub fn app() -> Result<Router> {
-    // Database and state setup
-    let db = ConnectionPool::new(":memory:", 16)?;
-    let cache = lru::LruCache::new(1000.try_into()?);
-
-    let state = Arc::new(AppState {
-        db: Box::new(db),
-        cache: Mutex::new(cache),
-    });
-
-    // CORS setup
-    let cors = CorsLayer::new()
-        .allow_origin(Any)
-        .allow_methods([Method::OPTIONS, Method::POST, Method::GET])
-        .allow_headers(Any)
-        .max_age(Duration::from_secs(60) * 60 * 24);
-
-    // Router setup
-    Ok(Router::new()
-        .route("/", get(handle_get).post(handle_post))
-        .with_state(state)
-        .layer(cors)
-        .layer(CompressionLayer::new())
-        .layer(TraceLayer::new_for_http()))
-}
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -90,7 +27,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         .init();
 
     // App setup
-    let app = app()?;
+    let app = app::app()?;
 
     // TLS configuration
     let mut config = RustlsConfig::from_pem_file(
